Tech-Demo Branch Initialized
Created a Tech-Demo branch that does the following:

-A fixation screen appears for 1 second
-The stimulus now shows up for 2 seconds
-Repeat

Switching scenes and implementing timers is fairly simple

To switch scenes you just use a single line of code that tells the engine to change between stimulus and fixation

The timer is a little complicated. What you do is create a method called a coroutine. This coroutine has a timer in it, and when the timer elapses it returns an input to the update function saying TimeElapsed. When update sees that TimeElapsed has returned it executes the scene change.

I also added in a script that checks to see if ESC has been pressed or not. If it's been pressed it terminates the program.

The il2cpp_cache has not been pushed due to file size limitations. These files are consistent across platforms, simply switch your build to WEBGL from the build menu to generate the WEBGL build options.